+++ setup.sh
@@ -28,33 +28,35 @@ rm -f $log
 
 #This colour
 cyan='\e[0;36m'
-green='\e[0;34m'
-okegreen='\033[92m'
+green='\e[0;32m'
 lightgreen='\e[1;32m'
 white='\e[1;37m'
 red='\e[1;31m'
 yellow='\e[1;33m'
-BlueF='\e[1;34m'
+blue='\e[1;34m'
 path=`pwd`
 
 #Check root dulu
 [[ `id -u` -eq 0 ]] || { echo -e $red "Must be root to run script"; exit 1; }
-resize -s 30 73 > /dev/null 2>&1
+resize -s 80 103 > /dev/null 2>&1
 clear
 
 #Banner dong biar keren
-echo -e $cyan ""
-echo "         [ ]=================================================[ ]";
-echo "         [ ]    ________       ____     __  ___       __     [ ]";
-echo "         [ ]     /_  __/ /  ___ / __/__ _/ /_/ _ \___ _/ /_  [ ]";
-echo "         [ ]     / / / _ \/ -_) _// _ \`/ __/ , _/ _ \`/ __/   [ ]";
-echo "         [ ]     /_/ /_//_/\__/_/  \_,_/\__/_/|_|\_,_/\__/   [ ]";
-echo "         [ ]=================================================[ ] ";
-echo "         [ ]          Setup.sh - configuration script        [ ]"
-echo "         [ ]        Use this script to configure fatrat      [ ]"
-echo "         [ ]              Install all dependencies           [ ]"
-echo "         [ ]=================================================[ ]";
+echo -e $green ""
+echo "___________         __  __________          __    "
+echo "\_   _____/_____  _/  |_\______   \_____  _/  |_  "
+echo " |    __)  \__  \ \   __\|       _/\__  \ \   __\ "
+echo " |     \    / __ \_|  |  |    |   \ / __ \_|  |   "
+echo " \___  /   (____  /|__|  |____|_  /(____  /|__|   "
+echo "     \/         \/              \/      \/        "
+echo "                 ____    ________                 "
+echo "                /_   |  /   __   \                "
+echo "                 |   |  \____    /                "
+echo "                 |   |     /    /                 "
+echo "                 |___| /\ /____/                  "
+echo "                       \/                         "
 echo ""
+echo -e $blue "         Setup Script for FATRAT 1.9       "
 touch $log
 echo "------------------------------------------------------" >> $log
 echo "| Tools paths configured in (setup.sh) for TheFatRat |" >> $log
@@ -63,48 +65,49 @@ echo "                                                      " >> $log
 #check if xterm is installed
 which xterm > /dev/null 2>&1
 if [ "$?" -eq "0" ]; then
-echo "[ ✔ ] Xterm.............................[ found ]"
+echo -e $green "[ ✔ ] Xterm.............................[ found ]"
 which xterm >> $log 2>&1
 sleep 2
 else
 echo ""
-echo "[ X ] xterm -> not found!                        ]"
-echo "[ ! ] This script requires xterm                 ]"
+echo -e $red "[ X ] xterm -> not found! "
 sleep 2
+echo -e $yellow "[ ! ] Installing Xterm                     "
+sleep 2
+echo -e $green ""
 sudo apt-get install xterm -y
 clear
-echo "[ ✔ ] Done installing .... "
+echo -e $green "[ ✔ ] Done installing .... "
 which xterm >> $log 2>&1
 fi
 
 #check if zenity its installed
 which zenity > /dev/null 2>&1
 if [ "$?" -eq "0" ]; then
-echo "[ ✔ ] Zenity............................[ found ]"
+echo -e $green "[ ✔ ] Zenity............................[ found ]"
 which zenity >> $log 2>&1
 sleep 2
 else
 echo ""
-echo "[ X ] zenity -> not found!                       ]"
-echo "[ ! ]  This script requires zenity               ]"
+echo -e $red "[ X ] zenity -> not found! "
 sleep 2
-echo "[ ! ]  Installing zenity from your apt sources   ]"
+echo -e $yellow "[ ! ]  Installing zenity                   "
 xterm -T "☣ INSTALL ZENITY ☣" -geometry 100x30 -e "sudo apt-get install zenity -y"
-echo "[ ✔ ] Done installing .... "
+echo -e $green "[ ✔ ] Done installing .... "
 which zenity >> $log 2>&1
 fi
 
 # check if gcc exists
 which gcc > /dev/null 2>&1
 if [ "$?" -eq "0" ]; then
-echo "[ ✔ ] Gcc compiler......................[ found ]"
+echo -e $green "[ ✔ ] Gcc compiler......................[ found ]"
 which gcc >> $log 2>&1
 sleep 2
 else
-echo "[ X ] gcc compiler      -> not found              ]"
-echo "[ ! ]   Installing gcc from your apt sources      ]"
+echo -e $red "[ X ] gcc compiler      -> not found "
+echo -e $yellow "[ ! ]   Installing gcc "
 xterm -T "☣ INSTALL GCC COMPILLER ☣" -geometry 100x30 -e "sudo apt-get install gcc -y"
-echo "[ ✔ ] Done installing .... "
+echo -e $green "[ ✔ ] Done installing .... "
 which gcc >> $log 2>&1
 sleep 2
 fi
@@ -112,14 +115,14 @@ fi
 # check if mingw32 exists
 which i586-mingw32msvc-gcc > /dev/null 2>&1
 if [ "$?" -eq "0" ]; then
-echo "[ ✔ ] Mingw32 Compiler..................[ found ]"
+echo -e $green "[ ✔ ] Mingw32 Compiler..................[ found ]"
 which i586-mingw32msvc-gcc >> $log 2>&1
 sleep 2
 else
-echo "[ X ] mingw32 compiler  -> not found               ]"
-echo "[ ! ]   Installing zenity from your apt sources    ]"
+echo -e $red "[ X ] mingw32 compiler  -> not found "
+echo -e $yellow "[ ! ]   Installing Mingw32 "
 xterm -T "☣ INSTALL MINGW32 COMPILLER ☣" -geometry 100x30 -e "sudo apt-get install mingw32 -y"
-echo "[ ✔ ] Done installing .... "
+echo -e $green "[ ✔ ] Done installing .... "
 which i586-mingw32msvc-gcc >> $log 2>&1
 sleep 2
 fi
@@ -127,44 +130,29 @@ fi
 # check if monodevelop exists
 which monodevelop > /dev/null 2>&1
 if [ "$?" -eq "0" ]; then
-echo "[ ✔ ] Monodevelop ......................[ found ]"
+echo -e $green "[ ✔ ] Monodevelop ......................[ found ]"
 which monodevelop >> $log 2>&1
 sleep 2
 else
-echo "[ X ] Monodevelop  -> not found                     ]"
-echo "[ ! ]  Installing monodevelop from your apt sources ]"
+echo -e $red "[ X ] Monodevelop  -> not found "
+echo -e $yellow "[ ! ]  Installing monodevelop "
 xterm -T "☣ INSTALL MONODEVELOP ☣" -geometry 100x30 -e "sudo apt-get install monodevelop -y"
-echo "[ ✔ ] Done installing ...."
+echo -e $green "[ ✔ ] Done installing ...."
 which monodevelop >> $log 2>&1
-sleep 2
-fi
-
-# check if ruby exists
-which ruby > /dev/null 2>&1
-if [ "$?" -eq "0" ]; then
-echo "[ ✔ ] Ruby .............................[ found ]"
-which ruby >> $log 2>&1
-sleep 2
-else
-echo "[ X ] Ruby  -> not found                             ]"
-echo "[ ! ]     Installing ruby from your apt sources      ]"
-xterm -T "☣ INSTALL RUBY ☣" -geometry 100x30 -e "sudo apt-get install ruby -y"
-echo "[ ✔ ] Done installing ...."
-which ruby >> $log 2>&1
 sleep 2
 fi
 
 #check if apache2 exists
 which apache2 > /dev/null 2>&1
 if [ "$?" -eq "0" ]; then
-echo "[ ✔ ] Apache2 ..........................[ found ]"
+echo -e $green "[ ✔ ] Apache2 ..........................[ found ]"
 which apache2 >> $log 2>&1
 sleep 2
 else
-echo "[ X ] Apache2 -> not found                            ]"
-echo "[ ! ]    Installing apache2 from your apt sources     ]"
+echo -e $red "[ X ] Apache2 -> not found  "
+echo -e $yellow "[ ! ]    Installing apache2 "
 xterm -T "☣ INSTALL APACHE2 ☣" -geometry 100x30 -e "sudo apt-get install apache2 -y"
-echo "[ ✔ ] Done installing ...."
+echo -e $green "[ ✔ ] Done installing ...."
 which apache2 >> $log 2>&1
 sleep 2
 fi
@@ -174,14 +162,14 @@ fi
 #gnome terminal is used in main script to run searchsploit
 which gnome-terminal > /dev/null 2>&1
 if [ "$?" -eq "0" ]; then
-echo "[ ✔ ] Gnome Terminal....................[ found ]"
+echo -e $green "[ ✔ ] Gnome Terminal....................[ found ]"
 which gnome-terminal >> $log 2>&1
 sleep 2
 else
-echo "[ X ] Gnome-terminal-> not found                      ]"
-echo "[ ! ] Installing gnome-terminal from your apt sources ]"
+echo -e $red "[ X ] Gnome-terminal-> not found "
+echo -e $yellow "[ ! ] Installing gnome-terminal "
 xterm -T "☣ INSTALL GNOME-TERMINAL ☣" -geometry 100x30 -e "sudo apt-get install gnome-terminal -y"
-echo "[ ✔ ] Done installing ...."
+echo -e $green "[ ✔ ] Done installing ...."
 which gnome-terminal >> $log 2>&1
 sleep 2
 fi
@@ -189,78 +177,97 @@ fi
 #Checking if upx compressor exists
 which upx > /dev/null 2>&1
 if [ "$?" -eq "0" ]; then
-echo "[ ✔ ] UPX Compressor....................[ found ]"
+echo -e $green "[ ✔ ] UPX Compressor....................[ found ]"
 which upx >> $log 2>&1
 sleep 2
 else
-echo "[ X ] Upx compressor  -> not found                    ]"
-echo "[ ! ] Installing upx-compressor from your apt sources ]"
+echo -e $red "[ X ] Upx compressor  -> not found "
+echo -e $yellow "[ ! ] Installing upx-compressor "
 xterm -T "☣ INSTALL UPX COMPRESSOR ☣" -geometry 100x30 -e "sudo apt-get install upx-ucl -y"
-echo "[ ✔ ] Done installing ...."
+echo -e $green "[ ✔ ] Done installing ...."
 which upx >> $log 2>&1
-sleep 2
-fi
-
-#Checking if keytool exists
-which keytool > /dev/null 2>&1
-if [ "$?" -eq "0" ]; then
-echo "[ ✔ ] Keytool (java)....................[ found ]"
-which keytool >> $log 2>&1
-sleep 2
-else
-echo "[ X ] Keytool (java) -> not found                    ]"
-echo "[ ! ] Installing Java from your apt sources ]"
-xterm -T "☣ INSTALL JAVA ☣" -geometry 100x30 -e "sudo apt-get install default-jre default-jdk -y "
-echo "[ ✔ ] Done installing ...."
-which keytool >> $log 2>&1
 sleep 2
 fi
 
 #Checking if Jarsigner exists
 which jarsigner > /dev/null 2>&1
 if [ "$?" -eq "0" ]; then
-echo "[ ✔ ] Jarsigner (java)..................[ found ]"
+echo -e $green "[ ✔ ] Jarsigner (java)..................[ found ]"
 which jarsigner >> $log 2>&1
+rm -f $config
+#Creating new config file 
+touch $config
+echo "********************************************************************************************************" >> $config
+echo "** Configuration Paths for TheFatRat , do not delete anything from this file or program will not work **" >> $config
+echo "**       if you need to reconfig your tools path , then run ./setup.sh in (TheFatRat directory) .     **" >> $config
+echo "********************************************************************************************************" >> $config
+echo "jarsigner" | tee -a $config >> /dev/null 2>&1
 sleep 2
 else
-echo "[ X ] Jarsigner (java) -> not found                    ]"
-echo "[ ! ] Installing Java from your apt sources ]"
-xterm -T "☣ INSTALL JAVA ☣" -geometry 100x30 -e "sudo apt-get install default-jdk -y "
-echo "[ ✔ ] Done installing ...."
+echo -e $red "[ X ] Jarsigner (java) -> not found "
+echo -e $yellow "[ ! ] Installing Java "
+xterm -T "☣ INSTALL OPENJDK-8 ☣" -geometry 100x30 -e "sudo apt-get install openjdk-8-jre openjdk-8-jdk -y "
+echo -e $green "[ ✔ ] Done installing ...."
 which jarsigner >> $log 2>&1
+rm -f $config
+#Creating new config file 
+touch $config
+echo "********************************************************************************************************" >> $config
+echo "** Configuration Paths for TheFatRat , do not delete anything from this file or program will not work **" >> $config
+echo "**       if you need to reconfig your tools path , then run ./setup.sh in (TheFatRat directory) .     **" >> $config
+echo "********************************************************************************************************" >> $config
+echo "jarsigner" | tee -a $config >> /dev/null 2>&1
 sleep 2
 fi
 
 #Checking if Unzip exists
 which unzip > /dev/null 2>&1
 if [ "$?" -eq "0" ]; then
-echo "[ ✔ ] Unzip.............................[ found ]"
+echo -e $green "[ ✔ ] Unzip.............................[ found ]"
 which unzip >> $log 2>&1
+echo "unzip" | tee -a $config >> /dev/null 2>&1
 sleep 2
 else
-echo "[ X ] Unzip -> not found                    ]"
-echo "[ ! ] Installing Unzip from your apt sources ]"
+echo -e $red "[ X ] Unzip -> not found "
+echo -e $yellow "[ ! ] Installing Unzip "
 xterm -T "☣ INSTALL UNZIP ☣" -geometry 100x30 -e "sudo apt-get install unzip -y "
-echo "[ ✔ ] Done installing ...."
+echo -e $green "[ ✔ ] Done installing ...."
 which unzip >> $log 2>&1
+echo "unzip" | tee -a $config >> /dev/null 2>&1
 sleep 2
 fi
 
-#Checking if Aapt exists
-which aapt > /dev/null 2>&1
+#Checking if keytool exists
+which keytool > /dev/null 2>&1
 if [ "$?" -eq "0" ]; then
-echo "[ ✔ ] Aapt..............................[ found ]"
-which aapt >> $log 2>&1
+echo -e $green "[ ✔ ] Keytool (java)....................[ found ]"
+which keytool >> $log 2>&1
+echo "keytool" | tee -a $config >> /dev/null 2>&1
 sleep 2
 else
-echo "[ X ] Aapt -> not found                    ]"
-echo "[ ! ] Installing Aapt from your apt sources ]"
-xterm -T "☣ INSTALL AAPT ☣" -geometry 100x30 -e "sudo apt-get install aapt -y "
-echo "[ ✔ ] Done installing ...."
-which aapt >> $log 2>&1
+echo -e $red "[ X ] Keytool (java) -> not found  "
+echo -e $yellow "[ ! ] Installing Java "
+xterm -T "☣ INSTALL JAVA ☣" -geometry 100x30 -e "sudo apt-get install openjdk-8-jdk -y "
+echo -e $green "[ ✔ ] Done installing ...."
+which keytool >> $log 2>&1
+echo "keytool" | tee -a $config >> /dev/null 2>&1
 sleep 2
 fi
 
+#Adding zipalign path to config
+echo -e $green "[ ✔ ] Zipalign "
+echo "$path/tools/android-sdk-25.0.2/zipalign" >> $log 2>&1
+echo "$path/tools/android-sdk-25.0.2/zipalign" | tee -a $config >> /dev/null 2>&1
+sleep 2
+
+
+#Adding Proguard path to config
+echo -e $green "[ ✔ ] Proguard "
+echo "$path/tools/proguard5.3.2/lib/proguard" >> $log 2>&1
+echo "$path/tools/proguard5.3.2/lib/proguard" | tee -a $config >> /dev/null 2>&1
+sleep 2
+
+xterm -T "☣ INSTALL APKCREATION DEPENDENCIES ☣" -geometry 100x30 -e "sudo apt-get install lib32z1 lib32ncurses5 lib32stdc++6 -y"
 #################################
 #inputrepo
 #################################
@@ -284,61 +291,62 @@ echo 'deb-src http://http.kali.org/kali kali-rolling main contrib non-free' >> /
 sleep 2
 xterm -T "☣ UPDATING KALI REPO ☣" -geometry 100x30 -e "sudo apt-get update" >>$log 2>&1
 
-#Checking if apktool exists
-which apktool > /dev/null 2>&1
+#Adding Dx & Aapt path to config 
+which aapt > /dev/null 2>&1
 if [ "$?" -eq "0" ]; then
-echo "[ ✔ ] Apktool...........................[ found ]"
-which apktool >> $log 2>&1
+echo -e $green "[ ✔ ] Dex2Jar...........................[ found ]"
+xterm -T "☣ Removing Your Current Aapt ☣" -geometry 100x30 -e "sudo apt-get remove --purge aapt -y" >>$log 2>&1
+unlink /usr/local/sbin/aapt > /dev/null 2>&1
+unlink /usr/bin/aapt > /dev/null 2>&1
+ln -s $path/tools/android-sdk-25.0.2/aapt /usr/local/sbin/aapt > /dev/null 2>&1
+echo -e $green "[ ✔ ] Dx & Aapt"
+echo "$path/tools/android-sdk-25.0.2/dx" >> $log 2>&1
+echo "$path/tools/android-sdk-25.0.2/dx" | tee -a $config >> /dev/null 2>&1
+echo "$path/tools/android-sdk-25.0.2/aapt" >> $log 2>&1
+echo "$path/tools/android-sdk-25.0.2/aapt" | tee -a $config >> /dev/null 2>&1
 sleep 2
 else
-echo "[ X ] Apktool  -> not found                     "
-echo "[ ! ] Installing apktool from Kali repositories "
-xterm -T "☣ INSTALL APKTOOOL ☣" -geometry 100x30 -e "sudo apt-get install apktool --force-yes -y"
-echo "[ ✔ ] Done installing ...."
-which apktool >> $log 2>&1
+echo -e $green "[ ✔ ] Dx & Aapt"
+echo "$path/tools/android-sdk-25.0.2/dx" >> $log 2>&1
+echo "$path/tools/android-sdk-25.0.2/dx" | tee -a $config >> /dev/null 2>&1
+echo "$path/tools/android-sdk-25.0.2/aapt" >> $log 2>&1
+echo "$path/tools/android-sdk-25.0.2/aapt" | tee -a $config >> /dev/null 2>&1
 sleep 2
 fi
+
+#Adding Apktool path to config
+echo -e $green "[ ✔ ] Apktool 2.2.2 "
+echo "$path/tools/apktool2.2.2/apktool" >> $log 2>&1
+echo "$path/tools/apktool2.2.2/apktool" | tee -a $config >> /dev/null 2>&1
+sleep 2
 
 #Checking if dex2jar exists
 which d2j-jar2dex > /dev/null 2>&1
 if [ "$?" -eq "0" ]; then
-echo "[ ✔ ] Dex2Jar...........................[ found ]"
+echo -e $green "[ ✔ ] Dex2Jar...........................[ found ]"
 which d2j-jar2dex >> $log 2>&1
+echo "d2j-jar2dex" | tee -a $config >> /dev/null 2>&1
 sleep 2
 else
-echo "[ X ] Dex2jar  -> not found                     "
-echo "[ ! ] Installing dex2jar from Kali repositories "
-xterm -T "☣ INSTALL APKTOOOL ☣" -geometry 100x30 -e "sudo apt-get install dex2jar --force-yes -y"
-echo "[ ✔ ] Done installing ...."
+echo -e $red "[ X ] Dex2jar  -> not found "
+echo -e $yellow "[ ! ] Installing dex2jar "
+xterm -T "☣ INSTALL DEX2JAR ☣" -geometry 100x30 -e "sudo apt-get install dex2jar --force-yes -y"
+echo -e $green "[ ✔ ] Done installing ...."
 which d2j-jar2dex >> $log 2>&1
+echo "d2j-jar2dex" | tee -a $config >> /dev/null 2>&1
 sleep 2
 fi
-
-#installing dependencies for ruby script
-echo "[ ! ] Installing dedepndencies for ruby script from Kali repositories "
-xterm -T "☣ INSTALL DEPENDENCIES ☣" -geometry 100x30 -e "sudo apt-get install libmagickwand-dev imagemagick --force-yes -y"
-echo "[ ✔ ] Done installing ...."
-sleep 2
 
 # check if metasploit-framework its installed
 which msfconsole > /dev/null 2>&1
 if [ "$?" -eq "0" ]; then
-echo -e $cyan "[ ✔ ] Metasploit-Framework..............[ found ]"
-# msf was detected , removing config file in case setup was already configured before
-rm -f $config
-
-#Creating new config file based on last detection of msf
-touch $config
-echo "********************************************************************************************************" >> $config
-echo "** Configuration Paths for TheFatRat , do not delete anything from this file or program will not work **" >> $config
-echo "**       if you need to reconfig your tools path , then run ./setup.sh in (TheFatRat directory) .     **" >> $config
-echo "********************************************************************************************************" >> $config
-echo "msfconsole" | tee -a $config $log > /dev/null 2>&1
-echo "msfvenom" | tee -a $config $log > /dev/null 2>&1
+echo -e $green "[ ✔ ] Metasploit-Framework..............[ found ]"
+echo "msfconsole" | tee -a $config $log >> /dev/null 2>&1
+echo "msfvenom" | tee -a $config $log >> /dev/null 2>&1
 sleep 2
 else
 echo ""
-echo -e $cyan "[ X ] metasploit-framework -> not found                         "
+echo -e $red "[ X ] metasploit-framework -> not found "
 
 # Providing manual input to user in case metasploit was installed from git and is not on system path
 
@@ -346,12 +354,7 @@ q1=$(zenity  --list  --radiolist  --column "Pick" --column "Action" TRUE "Setup 
 case $q1 in 
  
 "Setup Metasploit path manually")
-rm -f $config
-touch $config
-echo "********************************************************************************************************" >> $config
-echo "** Configuration Paths for TheFatRat , do not delete anything from this file or program will not work **" >> $config
-echo "**       if you need to reconfig your tools path , then run ./setup.sh in (TheFatRat directory) .     **" >> $config
-echo "********************************************************************************************************" >> $config
+
 minpm=$(zenity --entry --title="Metasploit Path Manual Input" --width=100 --height=100 --text="Write the location of your Metasploit Path?" --entry-text="/opt/metasploit-framework");
 ret=$?
 
@@ -373,26 +376,13 @@ fi
 ;;
 
 "Install Metasploit from Repository")
-echo -e $cyan "[ ! ] Installing Metasploit-Framework  "
+echo -e $yellow "[ ! ] Installing Metasploit-Framework  "
 xterm -T "☣ INSTALL METASPLOIT-FRAMEWORK ☣" -geometry 100x30 -e "sudo apt-get install metasploit-framework --force-yes -y"
-echo -e $cyan "[ ✔ ] Done installing ...."
-rm -f $config
-touch $config
-echo "********************************************************************************************************" >> $config
-echo "** Configuration Paths for TheFatRat , do not delete anything from this file or program will not work **" >> $config
-echo "**       if you need to reconfig your tools path , then run ./setup.sh in (TheFatRat directory) .     **" >> $config
-echo "********************************************************************************************************" >> $config
-# adding the msf startups automatically to config file
+echo -e $green "[ ✔ ] Done installing ...."
 echo "msfconsole" | tee -a $config $log > /dev/null 2>&1
 echo "msfvenom" | tee -a $config $log > /dev/null 2>&1
 ;;
 "Use default config")cd
-rm -f $config
-touch $config
-echo "********************************************************************************************************" >> $config
-echo "** Configuration Paths for TheFatRat , do not delete anything from this file or program will not work **" >> $config
-echo "**       if you need to reconfig your tools path , then run ./setup.sh in (TheFatRat directory) .     **" >> $config
-echo "********************************************************************************************************" >> $config
 echo "msfconsole" | tee -a $config $log > /dev/null 2>&1
 echo "msfvenom" | tee -a $config $log > /dev/null 2>&1
 ;;
@@ -402,11 +392,11 @@ fi
 
 which backdoor-factory > /dev/null 2>&1
 if [ "$?" -eq "0" ]; then
-echo -e $cyan "[ ✔ ] Backdoor-Factory..................[ found ]"
+echo -e $green "[ ✔ ] Backdoor-Factory..................[ found ]"
 echo "backdoor-factory" | tee -a $config $log > /dev/null 2>&1
 sleep 2
 else
-echo -e $cyan "[ X ] backdoor-factory  -> not found                  "
+echo -e $red "[ X ] backdoor-factory  -> not found "
 echo ""
 
 q2=$(zenity  --list  --radiolist  --column "Pick" --column "Action" TRUE "Setup Backdoor-Factory path manually" FALSE "Install Backdoor-Factory from Repository" FALSE "Use default config" --text="`printf "Backdoor-Factory was not detected in your system path ! \n Choose one of the options bellow ."`");
@@ -426,9 +416,9 @@ fi
 ;;
 
 "Install Backdoor-Factory from Repository")
-echo -e $cyan "[ ! ]   Installing backdoor-factory from kali repositories   ]"
+echo -e $yellow "[ ! ]   Installing backdoor-factory "
 xterm -T "☣ INSTALL BACKDOOR-FACTORY ☣" -geometry 100x30 -e "sudo apt-get install backdoor-factory --force-yes -y"
-echo -e $cyan "[ ✔ ] Done installing ...."
+echo -e $green "[ ✔ ] Done installing ...."
 echo "backdoor-factory" | tee -a $config $log > /dev/null 2>&1
 ;;
 
@@ -441,11 +431,11 @@ fi
 
 which searchsploit > /dev/null 2>&1
 if [ "$?" -eq "0" ]; then
-echo -e $cyan "[ ✔ ] Searchsploit......................[ found ]"
+echo -e $green "[ ✔ ] Searchsploit......................[ found ]"
 echo "searchsploit" | tee -a $config $log > /dev/null 2>&1
 sleep 2
 else
-echo -e $cyan "[ X ] searchsploit  -> not found]"
+echo -e $red "[ X ] searchsploit  -> not found"
 echo ""
 q3=$(zenity  --list  --radiolist  --column "Pick" --column "Action" TRUE "Setup Searchsploit path manually" FALSE "Install Searchsploit from Repository" FALSE "Use default config" --text="`printf "Searchsploit was not detected in your system path ! \n Choose one of the options bellow ."`");
 case $q3 in 
@@ -464,13 +454,13 @@ fi
 ;;
 
 "Install Searchsploit from Repository")
-echo -e $cyan "[ ! ]    Installing searchsploit from kali repositories      ]"
+echo -e $yellow "[ ! ]    Installing searchsploit "
 xterm -T "☣ INSTALL SEARCHSPLOIT ☣" -geometry 100x30 -e "sudo apt-get install exploitdb --force-yes -y"
-echo -e $cyan "[ ✔ ] Done installing ...."
+echo -e $green "[ ✔ ] Done installing ...."
 echo "searchsploit" | tee -a $config $log > /dev/null 2>&1
 sleep 2
 echo ""
-echo -e $cyan "Configuration and tool installed with success!";
+echo -e $green "Configuration and tool installed with success!";
 sleep 2
 ;;
 
@@ -483,7 +473,7 @@ fi
 ################################
 # rebackyo repo
 ################################
-echo "Reactivating you original repositories"
+echo -e $blue "Reactivating you original repositories"
 rm -f /etc/apt/sources.list
 mv /etc/apt/sources.list.backup /etc/apt/sources.list
 #now we can remove the emergency backup securely
@@ -512,8 +502,13 @@ exit
 fi
 if [ $lnk ==  "1" ];then
 chmod +x fatrat
+chmod +x $path/tools/android-sdk-25.0.2/zipalign
+chmod +x $path/tools/proguard5.3.2/lib/proguard
+chmod +x $path/tools/android-sdk-25.0.2/dx
+chmod +x $path/tools/android-sdk-25.0.2/aapt
+chmod +x $path/tools/apktool2.2.2/apktool
 zenity --width=100 --height=100 --no-wrap --info --text="To execute fatrat write in fatrat directory (./fatrat) to execute it."
-sleep2
+sleep 2
 echo -e $green "Instalation completed"
 fi
 exit